Move WorkState to module
