+++ HW04/ekaterina_balabanovich/lib/mentor.rb
@@ -13,18 +13,18 @@ class Mentor < Member
 
   def reject_to_work!(homework, student)
     solver = WorksManager.instance.find_solver(student, homework)
-    WorksManager.instance.transition_to(WorkStateRejected.new, solver, homework)
+    WorksManager.instance.transition_to(WorkState::Rejected.new, solver, homework)
   end
 
   def accept!(homework, student)
     solver = WorksManager.instance.find_solver(student, homework)
-    WorksManager.instance.transition_to(WorkStateAccepted.new, solver, homework)
+    WorksManager.instance.transition_to(WorkState::Accepted.new, solver, homework)
   end
 
   def assign(homework, student)
     WorksManager.instance.attach_solver(student, homework)
     solver = WorksManager.instance.find_solver(student, homework)
-    WorkStateNew.new.process(solver, homework)
+    WorkState::New.new.process(solver, homework)
   end
 
   def join_to_check(homework)
