+++ HW04/ekaterina_balabanovich/lib/work_state/base.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module WorkState
+  class Base
+    def process(members, homework); end
+  end
+end
