+++ HW04/ekaterina_balabanovich/lib/work_state/done.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module WorkState
+  class Done < Base
+    def process(members, homework)
+      WorksManager.instance.notify(members: members, task: homework, status: :done)
+      members.each { |a| a.get_homework(homework) unless a.homeworks.include?(homework) }
+    end
+  end
+end
