+++ HW04/ekaterina_balabanovich/lib/work_state/rejected.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module WorkState
+  class Rejected < Base
+    def process(members, homework)
+      WorksManager.instance.notify(members: members, task: homework, status: :reject)
+    end
+  end
+end
