--- HW04/ekaterina_balabanovich/lib/work_state/work_state.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-class WorkState
-  def process(members, homework); end
-end
