--- HW04/ekaterina_balabanovich/lib/work_state/work_state_rejected.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-class WorkStateRejected < WorkState
-  def process(members, homework)
-    WorksManager.instance.notify(members: members, task: homework, status: :reject)
-  end
-end
